% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anime.R
\name{anime}
\alias{anime}
\title{Match two sets of lines}
\usage{
anime(source, target, distance_tolerance = 10, angle_tolerance = 5)
}
\arguments{
\item{source}{a linestring geometry. Must be handleable by \code{wk}.}

\item{target}{a linestring geometry. Must be handleable by \code{wk}.}

\item{distance_tolerance}{the maximum distance between two linestrings to be considered a match.}

\item{angle_tolerance}{the maximum angle difference between two lines to be considered a match.}
}
\value{
an object of class \code{anime}
}
\description{
Match two sets of lines
}
