% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anime.R
\name{get_matches}
\alias{get_matches}
\title{Get the matches from an \code{anime} object}
\usage{
get_matches(x)
}
\arguments{
\item{x}{An \code{anime} object}
}
\description{
Get the matches from an \code{anime} object
}
