% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anime.R
\name{interpolate_intensive}
\alias{interpolate_intensive}
\title{Interpolate extensive variables
Interpolate values from the source geometry to the target geometry. Extensive properties are values which are dependent upon the geometry's size. Extensive properties would be a population or length.}
\usage{
interpolate_intensive(x, ...)
}
\arguments{
\item{x}{an \code{anime} object.}

\item{...}{variables to interpolate (must be numeric vectors).}
}
\description{
Interpolate extensive variables
Interpolate values from the source geometry to the target geometry. Extensive properties are values which are dependent upon the geometry's size. Extensive properties would be a population or length.
}
