% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{rnet_aggregate}
\alias{rnet_aggregate}
\title{Aggregate Network Matches}
\usage{
rnet_aggregate(
  source,
  matches,
  extensive_vars = NULL,
  intensive_vars = NULL,
  categorical_vars = NULL,
  source_len = sf::st_length(source),
  target_len = NULL
)
}
\arguments{
\item{source}{data.frame that contains attributes that will be attributed to the target dataset}

\item{matches}{results from \code{rnet_match()}. Contain \code{i}, \code{j}, and \code{shared_len}.}

\item{extensive_vars}{character vector of variables names in \code{source} that will be aggregated to the target as an extensive variable. See Details.}

\item{intensive_vars}{character vector of variable names in \code{source} that will be aggregated to hte target as intensive variables. See Details.}

\item{source_len}{default \code{sf::st_length(source)}. The length of linestrings in the source data set. Used for weighted extensive variables.}

\item{target_len}{the lengths of each line in the target dataset. Used for intensive variables.}
}
\description{
Aggregate Network Matches
}
