% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anime.R
\name{anime}
\alias{anime}
\title{Match two sets of lines}
\usage{
anime(source, target, distance_tolerance = 10, angle_tolerance = 5)
}
\arguments{
\item{source}{A geoarrow object containing the source lines}

\item{target}{A geoarrow object containing the target lines}

\item{distance_tolerance}{The maximum distance between two points to be considered a match}

\item{angle_tolerance}{The maximum angle between two lines to be considered a match}
}
\value{
A dataframe containing the matches
}
\description{
Match two sets of lines
}
