% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{anime_join}
\alias{anime_join}
\title{Join attributes from a source network to a target network}
\usage{
anime_join(
  x,
  y,
  distance_tolerance = 10,
  angle_tolerance = 5,
  suffix = c(".x", ".y"),
  match_strength = 0
)
}
\arguments{
\item{x}{An \code{sf} object or handleable geometry representing the target network.}

\item{y}{An \code{sf} object or handleable geometry representing the source network.}

\item{distance_tolerance}{The maximum distance between two linestrings to be considered a match.}

\item{angle_tolerance}{The maximum angle difference between two lines to be considered a match.}

\item{suffix}{A character vector of length 2 used to disambiguate non-joined duplicate variables.}

\item{match_strength}{A threshold for target_weighted to filter out weak matches (0 to 1). Default is 0.}
}
\value{
The target object with joined attributes.
}
\description{
Join attributes from a source network to a target network
}
