% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{anime_join}
\alias{anime_join}
\title{Join attributes from a source network to a destination network}
\usage{
anime_join(
  source,
  dest,
  min_overlap_dest = NULL,
  min_overlap_source = NULL,
  intensive = NULL,
  extensive = NULL,
  prefix = "",
  anime_opts = list()
)
}
\arguments{
\item{source}{An \code{sf} object or handleable geometry representing the source network.}

\item{dest}{An \code{sf} object or handleable geometry representing the destination network.}

\item{min_overlap_dest}{Optional. A scalar numeric threshold (0 to 1) for the destination segment overlap.}

\item{min_overlap_source}{Optional. A scalar numeric threshold (0 to 1) for the source segment overlap.}

\item{intensive}{Character vector of columns to treat as intensive (length-weighted average).}

\item{extensive}{Character vector of columns to treat as extensive (summed by source weight).}

\item{prefix}{A string prefix for the new columns. Default is "".}

\item{anime_opts}{A named list of matching options (e.g. \code{distance_tolerance} and \code{angle_tolerance}).}
}
\value{
The destination object with joined attributes.
}
\description{
Join attributes from a source network to a destination network
}
