% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anime.R
\name{filter_matches}
\alias{filter_matches}
\title{Filter matches in an anime object by target weighted strength}
\usage{
filter_matches(x, match_strength)
}
\arguments{
\item{x}{an \code{anime} object as created with \code{anime()}.}

\item{match_strength}{a threshold for filtering out weak matches.}
}
\value{
The modified \code{anime} object in-place (invisibly).
}
\description{
Filter matches in an anime object by target weighted strength
}
