% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anime.R
\name{interpolate_intensive}
\alias{interpolate_intensive}
\title{Interpolate extensive variables}
\usage{
interpolate_intensive(x, ...)
}
\arguments{
\item{x}{An \code{anime} object}

\item{...}{Variables to interpolate}
}
\description{
Interpolate extensive variables
}
